requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
chromadb>=0.5.0
fastembed>=0.3.0
httpx>=0.27.0
uvicorn>=0.29.0
starlette>=0.37.0
python-dotenv>=1.0.0
mcp>=1.0.0
# MLOps pipeline
apscheduler>=3.10.0
pandas>=2.0.0
